Revert "update popover"
