+++ frontend/src/Components/chatComponents/ChannelsRender.tsx
@@ -5,8 +5,6 @@ import channelStore from "../../store/channelStore";
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import PopoverMenu from "../common/Popover";
 import { io } from 'socket.io-client';
-import RenameChannel from "./RenameChannel";
-import DeleteChannel from "./DeleteChannel";
 
 interface Channel {
   id: string;
@@ -51,36 +49,21 @@ const ChannelsRender = ({ token }: { token: string }) => {
     setCurrentChannelPopover(el)
   };
 
+  const handleClosePopover = () => {
+    setAnchorEl(null);
+  };
+
   const open = Boolean(anchorEl);
 
   useEffect(() => {
     socket.on('newChannel', (newChannel) => {
       setNewChannel(newChannel)
     });
-
+  
     return () => {
       socket.off('newChannel')
     };
   },[setNewChannel])
-
-  const [openRenameChannel, setOpenRenameChannel] = useState(false);
-  const [openDeleteChannel, setOpenDeleteChannel] = useState(false);
-
-  const handleClosePopover = (action: String | null) => {
-    if (action === 'openRename') {
-      setOpenRenameChannel(true);
-    }
-    if (action === 'openDelete') {
-      setOpenDeleteChannel(true);
-    }
-    setAnchorEl(null);
-  };
-  const handleCloseRenameChannel = () => {
-    setOpenRenameChannel(false);
-  }
-  const handleCloseDeleteChannel = () => {
-    setOpenDeleteChannel(false);
-  }
 
   return (
     <>
@@ -132,8 +115,6 @@ const ChannelsRender = ({ token }: { token: string }) => {
               anchorEl={anchorEl}
               handleClosePopover={handleClosePopover}
             />
-            <RenameChannel open={openRenameChannel} handleClose ={handleCloseRenameChannel}/>
-            <DeleteChannel open={openDeleteChannel} handleClose ={handleCloseDeleteChannel} />
           </Box>
             }
           </ListItem>
