+++ frontend/src/store/channelStore.ts
@@ -9,7 +9,6 @@ interface Channel {
 type Store = {
   allChannels: Channel[];
   currentChannel: Channel;
-  defaultsChannel: Channel | null;
   currentChannelPopover: Channel;
   setCurrentChannel: (channel: Channel) => void;
   setAllChannels: (allChannels: Channel[]) => void;
@@ -22,15 +21,11 @@ const channelStore = create<Store>()(
     allChannels: [],
     currentChannel: {} as Channel,
     currentChannelPopover: {} as Channel,
-    defaultsChannel: null,
     setCurrentChannel: (currentChannel) => set({ currentChannel }),
     setAllChannels: (allChannels) => {
       set({ allChannels });
       if (get().allChannels.length > 0 && Object.keys(get().currentChannel).length === 0) {
-        set({
-          currentChannel: allChannels[0],
-          defaultsChannel: allChannels[0],
-        });
+        set({ currentChannel: get().allChannels[0] });
       }},
     setNewChannel: (newChannel) => set({ allChannels: [...get().allChannels, newChannel] }),
     setCurrentChannelPopover: (currentChannelPopover) => set({ currentChannelPopover }),
